Refactor decorator to loadable + mediator
