+++ Demo/BaseViewController.swift
@@ -77,7 +77,7 @@ class BaseCollectionViewController<T: CollectionViewSource>: BaseViewController<
   }
 }
 
-class BaseLoaderViewController<T: UIView>: BaseViewController<T>, Loader where T: AccessorView, T.Source: LoaderReusableSource {
+class BaseLoaderViewController<T: UIView>: BaseViewController<T>, Loadable where T: AccessorView, T.Source: LoaderReusableSource {
 
   convenience init(type: LoaderResult) {
     self.init()
@@ -165,7 +165,7 @@ class BaseLoaderViewController<T: UIView>: BaseViewController<T>, Loader where T
     containerView.source.disappear()
   }
 
-  func performLoading(intent: LoaderIntent) -> SectionObservable? {
+  func load(for intent: LoaderIntent) -> Observable<[Sectionable]?>? {
     var currentPage = 1
     var result: [Sectionable]? = nil
     switch intent {
